+++ Makefile
@@ -54,10 +54,10 @@ CPPFLAGS += -MD -MP -Isrc/common -DUSE_SSE4
 LDFLAGS  += -llzma -lpthread -lrt
 
 TARGETS         := $(addprefix bin/, aobaz autousi server playshogi crc64 extract net-test gencode ocldevs)
-AUTOUSI_BASES   := $(addprefix src/autousi/, autousi client play) $(addprefix src/common/, iobase option jqueue xzi err shogibase osi child nnet nnet-cpu nnet-ocl nnet-ipc opencl)
+AUTOUSI_BASES   := $(addprefix src/autousi/, autousi client play) $(addprefix src/common/, iobase option jqueue xzi err shogibase osi child nnet nnet-cpu nnet-ocl nnet-srv opencl)
 SERVER_BASES    := $(addprefix src/server/, server listen datakeep logging) $(addprefix src/common/, iobase xzi jqueue err option shogibase osi)
 GENCODE_BASES   := src/gencode/gencode
-PLAYSHOGI_BASES := src/playshogi/playshogi $(addprefix src/common/, option err iobase xzi shogibase osi child nnet nnet-cpu nnet-ocl nnet-ipc opencl)
+PLAYSHOGI_BASES := src/playshogi/playshogi $(addprefix src/common/, option err iobase xzi shogibase osi child nnet nnet-cpu nnet-ocl nnet-srv opencl)
 CRC64_BASES     := src/crc64/crc64 $(addprefix src/common/, xzi err iobase osi)
 EXTRACT_BASES   := src/extract/extract $(addprefix src/common/, xzi err iobase osi)
 OCLDEVS_BASES   := src/ocldevs/ocldevs src/common/err src/common/opencl
