+++ Makefile.config
@@ -5,13 +5,13 @@
 # ENABLE_ASSERT_AOBA := 1
 
 # Uncomment if you want to use OpenCL
-USE_OpenCL_AOBA := 1
+# USE_OpenCL_AOBA := 1
 
 # Uncomment if you want to use OpenMP
-# USE_OpenMP_AOBA := 1
+USE_OpenMP_AOBA := 1
 
 # Specify one of IntelMKL and OpenBLAS
-# BLAS_AOBA := OpenBLAS
+BLAS_AOBA := OpenBLAS
 
 # Uncomment if you need to specify path to OpenBLAS
 # OpenBLAS_INC_AOBA := /opt/OpenBLAS/include
