+++ readme.txt
@@ -94,10 +94,13 @@ Visual Studio 2017の「VS2017用 x64 Native Toolsコマンドプロンプト」
 
 - aobaz.exe のビルド
 
-Visual Studio 2017 で src/usi_engine/msvc/aoba-zero2017.sln を開きビルドします。
-「Windows SDK バージョン 10.0.17763.0 が見つかりません」のエラーメッセージが出
-る場合はプロジェクトプロパティの「全般」カテゴリの WindowsSDK バージョンを適切
-なバージョンに変更します。
+まず、上述の autousi.exe のビルドを済ませてください。これによって生成される
+幾つかのインクルードファイルが必要になります。そして、Visual Studio 2017
+src/usi_engine/msvc/aoba-zero2017.sln を開き、構成マネージャーで構成を
+Release、プラットフォームを x64 に設定しビルドします。「Windows SDK バージョ
+ン 10.0.17763.0 が見つかりません」のエラーメッセージが出る場合はプロジェクト
+プロパティの「全般」カテゴリの WindowsSDK バージョンを適切なバージョンに変更
+します。
 
 もしCPUのみで計算を行うプログラムをビルドしたいならば、プロジェクトプロパティ
 の「C/C++、プリプロセッサ」カテゴリのプリプロセッサの定義で「;USE_CPU_ONLY」を
