+++ src/autousi/play.cpp
@@ -6,7 +6,7 @@
 #include "child.hpp"
 #include "err.hpp"
 #include "iobase.hpp"
-#include "nnet-ipc.hpp"
+#include "nnet-srv.hpp"
 #include "osi.hpp"
 #include "param.hpp"
 #include "play.hpp"
