+++ src/common/nnet.hpp
@@ -1,6 +1,7 @@
 // 2019 Team AobaZero
 // This source code is in the public domain.
 #pragma once
+#include "iobase.hpp"
 #include "shogibase.hpp"
 #include <memory>
 #include <vector>
@@ -9,6 +10,8 @@ namespace NNAux {
   using uint   = unsigned int;
   using row_t  = std::unique_ptr<float []>;
   using wght_t = std::vector<std::pair<uint, row_t>>;
+  constexpr uint maxnum_nipc    = 128U;
+  constexpr uint maxnum_nnet    = 64U;
   constexpr uint width          = 9U;
   constexpr uint height         = 9U;
   constexpr uint size_plane     = width * height;
@@ -38,3 +41,21 @@ public:
   void take_action(const Action &a) noexcept;
   void encode_input(float *p) const noexcept;
 };
+
+enum class SrvType : uint { Register, FeedForward, FlushON, FlushOFF,
+    NNReset, End };
+
+struct SharedService {
+  uint id_ipc_next;
+  FName fn_weights;
+  uint njob;
+  struct { uint id; SrvType type; } jobs[NNAux::maxnum_nipc + 16U];
+};
+
+struct SharedIPC {
+  uint nnet_id;
+  float input[NNAux::size_input];
+  uint size_nnmove;
+  ushort nnmoves[SAux::maxsize_moves];
+  float probs[SAux::maxsize_moves];
+  float value; };
